main page scrolling partially implemented
it is now possible to scroll through images of a single artist. the next step is to make it possible to change artists. this should be easy enough, as you just need to increase the artistIndex by clicking buttons. but at the same time you must implement functionality, so that the user can register which artists he likes